Updated key to be a secret and moved all dependencies to their respective files
