--- Synview-backend/deps.ts
@@ -1,10 +1,0 @@
-export { withAccelerate } from "@prisma/extension-accelerate";
-export { PrismaClient } from './generated/prisma/client.ts';
-export { Application, Router, Context } from "@oak/oak"
-export {create, verify, decode, getNumericDate} from "@zaubrik/djwt";
-export { hash, verify as bycryptVerify } from "@felix/bcrypt"
-export { Session } from "https://deno.land/x/oak_sessions/mod.ts";
-export { oakCors } from "https://deno.land/x/cors/mod.ts";
-export { z } from 'zod'
-
-
