+++ Synview-backend/utils/JWTHelpers.ts
@@ -1,8 +1,16 @@
+import { Context } from "@oak/oak";
+import { AppState } from "../../common/types.ts";
+import { decode } from "@zaubrik/djwt";
+
+const key = Deno.env.get("AUTH_KEY");
+const keyBytes = new TextEncoder().encode(key)
 export async function generateKey() {
-  return await crypto.subtle.generateKey(
+  return await crypto.subtle.importKey(
+    "raw",
+    keyBytes,
     { name: "HMAC", hash: "SHA-512" },
     true,
-    ["sign", "verify"]
+    ["sign", "verify"],
   );
 }
 
@@ -22,4 +30,26 @@ export function getToken(auth: string) {
   }
 
   return token;
+}
+
+export function getPayloadFromToken(context: Context<AppState>) {
+  try {
+    const auth = context.state.session.get("Authorization");
+    if (!auth) {
+      return null;
+    }
+    const token = getToken(String(auth));
+    if (!token) {
+      return null;
+    }
+
+    const [a, payload, b] = decode(token); // need to import all three because if not i dont get payload correctly
+    if (!payload) {
+      throw new Error("Couldn't get payload");
+    }
+    return payload;
+  } catch (error) {
+    context.throw(400, "Error in token" + error);
+    return null;
+  }
 }
