+++ common/types.ts
@@ -1,3 +1,4 @@
+import { Session } from "https://deno.land/x/oak_sessions/mod.ts";
 import {
   UserPayloadSchema,
   UserInfoSchema,
@@ -8,11 +9,14 @@ import {
 } from "./schemas.ts";
 import { z } from "npm:zod";
 
+export type AppState = { session: Session };
 export type Projects = Project[];
-export type PostProject= z.infer<typeof PostProjectSchema>
-export type Project = z.infer<typeof ProjectSchema>
-export type LoginResponse = {token? : string}
+export type PostProject = z.infer<typeof PostProjectSchema>;
+export type Project = z.infer<typeof ProjectSchema>;
+export type LoginResponse = { token?: string };
 export type UserPayload = z.infer<typeof UserPayloadSchema>;
 export type UserInfo = z.infer<typeof UserInfoSchema>;
-export type EmailLoginRequestSchema = z.infer<typeof EmailLoginRequestSchema>
-export type EmailRegisterRequestSchema = z.infer<typeof EmailRegisterRequestSchema>
+export type EmailLoginRequestSchema = z.infer<typeof EmailLoginRequestSchema>;
+export type EmailRegisterRequestSchema = z.infer<
+  typeof EmailRegisterRequestSchema
+>;
